Move firebase utils into src
