+++ src/components/TraditionalDevotionals.tsx
@@ -1,7 +1,7 @@
-import { auth, favoriteTradDevo, unfavoriteTradDevo } from "@root/firebase";
 import { TTradDevo } from "@root/types";
 import { Container } from "@src/components/Container";
 import { DevotionalCard } from "@src/components/DevotionalCard";
+import { auth, favoriteTradDevo, unfavoriteTradDevo } from "@src/firebase";
 import { useTradDevos } from "@src/hooks/useTradDevos";
 import useStore from "@src/Store";
 import React, { useEffect, useState } from "react";
