+++ src/hooks/useSermons.ts
@@ -1,5 +1,5 @@
-import { db } from "@root/firebase";
 import { TSermon } from "@root/types";
+import { db } from "@src/firebase";
 import useStore from "@src/Store";
 import { collection, getDocs, orderBy, query } from "firebase/firestore";
 import { useEffect, useState } from "react";
