+++ src/screens/SermonsScreen.tsx
@@ -1,7 +1,7 @@
-import { auth, favoriteSermon, unfavoriteSermon } from "@root/firebase";
 import { TSermon } from "@root/types";
 import { Container } from "@src/components/Container";
 import { Sermon } from "@src/components/Sermon";
+import { auth, favoriteSermon, unfavoriteSermon } from "@src/firebase";
 import { useAudio } from "@src/hooks/useAudio";
 import { useSermons } from "@src/hooks/useSermons";
 import useStore, { useAudioStore } from "@src/Store";
